ContentCollector
Base concept, tests

+++ otp-alap/wsj-receiver/src/main/java/hu/farago/wsj/model/dao/mongo/ArticleCollectionManager.java
@@ -1,10 +1,12 @@
 package hu.farago.wsj.model.dao.mongo;
 
+import java.math.BigInteger;
+
 import hu.farago.wsj.model.entity.mongo.ArticleCollection;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ArticleCollectionManager extends MongoRepository<ArticleCollection, Long> {
+public interface ArticleCollectionManager extends MongoRepository<ArticleCollection, BigInteger> {
 	
 	ArticleCollection findByUrl(String lastname);
